refactor: 사이드바 폴더구조 변경
