+++ src/app/dashboard/[id]/page.tsx
@@ -1,0 +1,1 @@
+export default function Page() {}
