+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Sidebar from "@/src/components/common/sidebar";
+import Sidebar from "@/src/components/layout/sidebar";
 
 export default function Home() {
   return <Sidebar />;
